Major improvements.
CStruct creator revamped, simpler to setup.
ObjectToData, DataToData now output to std::string.
Added PropertyUInt32
Finalized PropertyStdVector (better design, export capable)
Improved build for 64 bit targets.
Added ToFile, FromFile string helpers.

+++ Private/Cacus.cpp
@@ -12,7 +12,7 @@ int32 volatile COpenThreads = 0;
 
 
 
-void CFixedArray::Setup( uint32 NewNum, size_t ElementSize)
+void CFixedArray::Setup( size_t NewNum, size_t ElementSize)
 {
 	if ( NewNum != Num )
 	{
